backend/pyproject.toml
[tool.poetry]
name = "backend"
version = "0.1.0"
description = "Auth Backend for cube-project"
authors = ["Shivam Singhal <shivam.singhal212@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.104.1"
uvicorn = "^0.24.0.post1"
pyjwt = "^2.8.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
